+++ src/main/java/ek/dfofb/projektkalkulationsvaerktoej/controller/TaskController.java
@@ -204,7 +204,8 @@ public class TaskController {
             return "redirect:/account/login";
         }
         if (httpSession.getAttribute("currentProject") == null || httpSession.getAttribute("currentTask") == null ||
-                !taskService.getTaskByID((Integer) httpSession.getAttribute("currentTask")).getName().equalsIgnoreCase(taskName)) {
+                !taskService.getTaskByID((Integer) httpSession.getAttribute("currentTask")).getName().equalsIgnoreCase(taskName) ||
+                !taskService.getAllAccountsAssignedToTask((Integer) httpSession.getAttribute("currentTask")).contains(account)) {
             return "redirect:/project/list";
         }
         Task task = taskService.getTaskByID((Integer) httpSession.getAttribute("currentTask"));
@@ -234,8 +235,12 @@ public class TaskController {
 
     @PostMapping("/task/complete")
     public String markTaskAsDone(@ModelAttribute Task task, HttpSession httpSession) {
-        if (httpSession.getAttribute("account") == null) {
+        Account account = (Account) httpSession.getAttribute("account");
+        if (account == null) {
             return "redirect:/account/login";
+        }
+        if (!taskService.getAllAccountsAssignedToTask((Integer) httpSession.getAttribute("currentTask")).contains(account)) {
+            return "redirect:/project";
         }
         taskService.markAsDone(task.getTaskID(), task.getHoursSpentOnTask());
         String taskName = taskService.getTaskByID(task.getTaskID()).getName();
@@ -270,6 +275,24 @@ public class TaskController {
         return saveCurrentTaskID(task.getName(), task.getTaskID(), httpSession);
     }
 
+    @GetMapping("/{projectName}/{taskName}/delete")
+    public String showDeleteForm(Model model, @PathVariable String projectName, @PathVariable String taskName, HttpSession httpSession) {
+        Account account = (Account) httpSession.getAttribute("account");
+        if (account == null) {
+            return "redirect:/account/login";
+        }
+        if (!authorizationService.hasPermission(account.getRoleID(), Permission.DELETE_TASKS)) {
+            return "redirect:/project";
+        }
+        if (httpSession.getAttribute("currentProject") == null || httpSession.getAttribute("currentTask") == null ||
+                !taskService.getTaskByID((Integer) httpSession.getAttribute("currentTask")).getName().equalsIgnoreCase(taskName)) {
+            return "redirect:/project/list";
+        }
+        Task task = taskService.getTaskByID((Integer) httpSession.getAttribute("currentTask"));
+        model.addAttribute("task", task);
+        return "delete-task-form";
+    }
+
     @PostMapping("/task/delete")
     public String deleteTask(@ModelAttribute Task task, HttpSession httpSession) {
         Account myAccount = (Account) httpSession.getAttribute("account");
@@ -280,10 +303,16 @@ public class TaskController {
             return "redirect:/project";
         }
         if (taskService.getAllAccountsAssignedToTask(task.getTaskID()).contains(myAccount) &&
-        taskService.getAllSubTasks(task.getTaskID()).isEmpty()) {
-            taskService.deleteTask(task.getTaskID());
+                taskService.getAllSubTasks(task.getTaskID()).isEmpty()) {
         }
-        return "redirect:/project";
+
+        if (taskService.getTaskByID(task.getTaskID()).getName().equalsIgnoreCase(task.getName())) {
+            taskService.deleteTask(task.getTaskID());
+            return "redirect:/project";
+        }
+        return "redirect:/project/" +
+                projectService.getProjectByID((Integer) httpSession.getAttribute("currentProject")).getName()
+                + "/" + taskService.getTaskByID(task.getTaskID()).getName() + "/delete";
     }
 
 }
